+++ src/modules/external/mci/components/types.ts
@@ -1,0 +1,6 @@
+import { DynamicInputCommonProps } from '@/modules/form/types';
+
+export interface MciClearanceProps extends DynamicInputCommonProps {
+  value?: string | null;
+  onChange: (value?: string | null) => void;
+}
